requirements_0-7-2.txt
vllm==0.7.2
transformers==4.50.0
datasets==3.3.2
huggingface_hub==0.30.2
accelerate==0.26.0
igraph==0.11.8
ipdb
nltk==3.9.1
pyserini==0.44.0
rank-bm25>=0.2.2
tqdm>=4.64.0
faiss-gpu==1.7.2
google-genai
numpy==1.26.4